backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
websockets
openai
httpx
browser-use-sdk
firecrawl-py
dnspython
langfuse
